Fix incorrect TCPv4Transport cached connection handling
Using the file descriptor (FD) as the cache key for a connection led to
the incorrect handling of sockets. This is because a FD may be re-used
by a different socket if the "old" socket is closed.

This diff uses `id(socket)` for the cache key, and cleans up the usage
of the cache and how connections are closed.

It also more reliably identifies sockets that need to be closed before
passing them from `get_next_connection` to `receive_message`.

It also fixes dropped events (from a wrongfully included `break` in
`_get_next_connection`).